refactor: remove miradorAnnotationPlugin container as its not used
